tried to fix dt too small

+++ src/game/game_util.h
@@ -22,8 +22,8 @@ extern TileArray tiles;
 extern WallArray walls;
 extern Player player;
 
-extern f64 game_time;
-extern f64 game_dt;
+extern f32 game_time;
+extern f32 game_dt;
 extern bool game_paused;
 
 #endif
